review coments#2
