+++ uniprotkb-rest/src/test/java/org/uniprot/api/uniprotkb/controller/UniProtKBGroupByControllerIT.java
@@ -1,0 +1,56 @@
+package org.uniprot.api.uniprotkb.controller;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.test.web.servlet.MockMvc;
+
+import static org.hamcrest.Matchers.containsStringIgnoringCase;
+import static org.hamcrest.core.Is.is;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.log;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+
+public abstract class UniProtKBGroupByControllerIT {
+    protected static final String EMPTY_PARENT = "";
+    private static final String INVALID_ORGANISM_ID = "36";
+
+
+    @Test
+    void getGroupByKeyword_emptyResults() throws Exception {
+        prepareSingleRootNodeWithNoChildren();
+
+        getMockMvc().perform(
+                        get(getPath())
+                                .param("query", "organism_id:" + INVALID_ORGANISM_ID)
+                                .param("parent", EMPTY_PARENT))
+                .andDo(log())
+                .andExpect(jsonPath("$.groups.size()", is(0)))
+                .andExpect(jsonPath("$.ancestors.size()", is(0)));
+    }
+
+    @Test
+    void getGroupByKeyword_whenFreeFormQueryAndEmptyResults() throws Exception {
+        prepareSingleRootNodeWithNoChildren();
+
+        getMockMvc().perform(get(getPath()).param("query", INVALID_ORGANISM_ID).param("parent", EMPTY_PARENT))
+                .andDo(log())
+                .andExpect(jsonPath("$.groups.size()", is(0)))
+                .andExpect(jsonPath("$.ancestors.size()", is(0)));
+    }
+
+    @Test
+    void getGroupByKeyword_whenQueryNotSpecified() throws Exception {
+        getMockMvc().perform(get(getPath()).param("parent", EMPTY_PARENT))
+                .andDo(log())
+                .andExpect(status().isBadRequest())
+                .andExpect(
+                        content()
+                                .string(
+                                        containsStringIgnoringCase(
+                                                "query is a required parameter")));
+    }
+
+    protected abstract MockMvc getMockMvc();
+    protected abstract String getPath();
+
+    protected abstract void prepareSingleRootNodeWithNoChildren() throws Exception;
+}
